% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer64.R
\name{allNA}
\alias{allNA}
\title{Not Available / Missing Values}
\usage{
allNA(x)
}
\arguments{
\item{x}{An R object to be tested.}
}
\description{
The function allNA implements all(is.na(x)) in a possibly faster way for integer64
}
