% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitwise64.R
\name{bitwise}
\alias{bitwise}
\alias{bitwNot}
\alias{bitwAnd}
\alias{bitwOr}
\alias{bitwXor}
\alias{bitwShiftL}
\alias{bitwShiftR}
\title{Bitwise Logical Operations}
\usage{
bitwNot(a)

bitwAnd(a, b)

bitwOr(a, b)

bitwXor(a, b)

bitwShiftL(a, n)

bitwShiftR(a, n)
}
\arguments{
\item{a, b}{integer vectors; numeric vectors are coerced to integer vectors.}

\item{n}{non-negative integer vector of values up to 31.}
}
\value{
An integer64 vector of length the longer of the arguments, or zero length if one is zero-length.
}
\description{
Logical operations on integer vectors with elements viewed as sets of bits. As soon as an integer64 vector is involved, the operations are performed using integer64 semantics. Otherwise the \code{base} package functions are called.
}
\examples{
x <- as.integer64(1:5)
y <- c(1L, 3L, 5L, 7L)
bitwAnd(x, y)
bitwOr(x, y)
bitwXor(x, y)
bitwNot(x)
bitwShiftL(x, 1L)
bitwShiftR(x, 1L)

}
\seealso{
\link[base:bitwAnd]{bitwAnd}, \link[base:bitwOr]{bitwOr}, \link[base:bitwXor]{bitwXor}, \link[base:bitwNot]{bitwNot}, \link[base:bitwShiftL]{bitwShiftL}, \link[base:bitwShiftR]{bitwShiftR}
}
