% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer64.R
\name{extract.replace.integer64}
\alias{extract.replace.integer64}
\alias{[.integer64}
\alias{[<-.integer64}
\alias{[[.integer64}
\alias{[[<-.integer64}
\title{Extract or Replace Parts of an integer64 vector}
\usage{
\method{[}{integer64}(x, i, j, ..., drop = TRUE)

\method{[}{integer64}(x, ...) <- value

\method{[[}{integer64}(x, ...)

\method{[[}{integer64}(x, ...) <- value
}
\arguments{
\item{x}{an atomic vector}

\item{i, j}{indices specifying elements to extract}

\item{drop}{relevant for matrices and arrays. If TRUE the result is coerced to the lowest possible dimension.}

\item{...}{further arguments to the \code{\link[=NextMethod]{NextMethod()}}}

\item{value}{an atomic vector with values to be assigned}
}
\value{
A vector, matrix, array or scalar of class 'integer64'
}
\description{
Methods to extract and replace parts of an integer64 vector.
}
\examples{
  as.integer64(1:12)[1:3]
  x <- matrix(as.integer64(1:12), nrow = 3L)
  x
  x[]
  x[,2:3]
}
\seealso{
\code{\link[base:Extract]{[}} \code{\link[=integer64]{integer64()}}
}
\keyword{classes}
\keyword{manip}
