% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer64.R
\name{identical.integer64}
\alias{identical.integer64}
\title{Identity function for class 'integer64'}
\arguments{
\item{x}{atomic vector of class 'integer64'}

\item{y}{atomic vector of class 'integer64'}

\item{num.eq}{see [identical()]}

\item{single.NA}{see [identical()]}

\item{attrib.as.set}{see [identical()]}

\item{ignore.bytecode}{see [identical()]}

\item{ignore.environment}{see [identical()]}

\item{ignore.srcref}{see [identical()]}

\item{extptr.as.ref}{see [identical()]}
}
\value{
A single logical value, `TRUE` or `FALSE`, never `NA` and never
  anything other than a single value.
}
\description{
This will discover any deviation between objects containing integer64 vectors.
}
\details{
This is simply a wrapper to [identical()] with default arguments
  `num.eq = FALSE, single.NA = FALSE`.
}
\examples{
  i64 <- as.double(NA); class(i64) <- "integer64"
  identical(i64-1, i64+1)
  identical.integer64(i64-1, i64+1)
}
\seealso{
[==.integer64] [identical()] [integer64()]
}
\keyword{classes}
\keyword{manip}
