% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer64.R
\name{identical.integer64}
\alias{identical.integer64}
\title{Identity function for class 'integer64'}
\arguments{
\item{x}{atomic vector of class 'integer64'}

\item{y}{atomic vector of class 'integer64'}

\item{num.eq}{see \code{\link[=identical]{identical()}}}

\item{single.NA}{see \code{\link[=identical]{identical()}}}

\item{attrib.as.set}{see \code{\link[=identical]{identical()}}}

\item{ignore.bytecode}{see \code{\link[=identical]{identical()}}}

\item{ignore.environment}{see \code{\link[=identical]{identical()}}}

\item{ignore.srcref}{see \code{\link[=identical]{identical()}}}

\item{extptr.as.ref}{see \code{\link[=identical]{identical()}}}
}
\value{
A single logical value, \code{TRUE} or \code{FALSE}, never \code{NA} and never
anything other than a single value.
}
\description{
This will discover any deviation between objects containing integer64 vectors.
}
\details{
This is simply a wrapper to \code{\link[=identical]{identical()}} with default arguments
\verb{num.eq = FALSE, single.NA = FALSE}.
}
\examples{
  i64 <- as.double(NA); class(i64) <- "integer64"
  identical(i64-1, i64+1)
  identical.integer64(i64-1, i64+1)
}
\seealso{
\link{==.integer64} \code{\link[=identical]{identical()}} \code{\link[=integer64]{integer64()}}
}
\keyword{classes}
\keyword{manip}
