% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops64.R
\name{ops64}
\alias{ops64}
\alias{binattr}
\alias{+.integer64}
\alias{-.integer64}
\alias{\%/\%.integer64}
\alias{\%\%.integer64}
\alias{*.integer64}
\alias{^.integer64}
\alias{/.integer64}
\alias{==.integer64}
\alias{!=.integer64}
\alias{<.integer64}
\alias{<=.integer64}
\alias{>.integer64}
\alias{>=.integer64}
\alias{&.integer64}
\alias{|.integer64}
\alias{!.integer64}
\title{Binary operators for integer64 vectors}
\usage{
binattr(e1, e2)

\method{+}{integer64}(e1, e2)

\method{-}{integer64}(e1, e2)

\method{\%/\%}{integer64}(e1, e2)

\method{\%\%}{integer64}(e1, e2)

\method{*}{integer64}(e1, e2)

\method{^}{integer64}(e1, e2)

\method{/}{integer64}(e1, e2)

\method{==}{integer64}(e1, e2)

\method{!=}{integer64}(e1, e2)

\method{<}{integer64}(e1, e2)

\method{<=}{integer64}(e1, e2)

\method{>}{integer64}(e1, e2)

\method{>=}{integer64}(e1, e2)

\method{&}{integer64}(e1, e2)

\method{|}{integer64}(e1, e2)

\method{!}{integer64}(x)
}
\arguments{
\item{e1, e2, x}{numeric or complex vectors or objects which can be coerced to such,
or other objects for which methods have been written for - especially 'integer64' vectors.}
}
\value{
\itemize{
\item \code{\link{&}}, \code{\link{|}}, \code{\link{!}}, \code{\link{!=}}, \code{\link{==}}, \code{\link{<}}, \code{\link{<=}}, \code{\link{>}}, \code{\link{>=}} return a logical vector
\item \code{\link{/}} returns a double vector
\item \code{\link{+}}, \code{\link{-}}, \code{\link{*}}, \code{\link[=Arithmetic]{\%/\%}}, \code{\link[=Arithmetic]{\%\%}}, \code{\link{^}} return a vector of class
'integer64' or different class depending on the operands
}
}
\description{
Binary operators for integer64 vectors.
}
\examples{
  as.integer64(1:12) - 1
  options(integer64_semantics="new")
  d <- 2.5
  i <- as.integer64(5)
  d/i  # new 0.5
  d*i  # new 13
  i*d  # new 13
  options(integer64_semantics="old")
  d/i  # old: 0.4
  d*i  # old: 10
  i*d  # old: 13
}
\seealso{
\code{\link[=integer64]{integer64()}}
}
\keyword{classes}
\keyword{manip}
