% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlevel64.R
\name{prank}
\alias{prank}
\title{(P)ercent (Rank)s}
\usage{
prank(x, ...)
}
\arguments{
\item{x}{a integer64 vector}

\item{...}{ignored}

\item{method}{NULL for automatic method selection or a suitable low-level
method, see details}
}
\value{
`prank` returns a numeric vector of the same length as `x`.
}
\description{
Function `prank.integer64`  projects the values `[min..max]` via ranks
  `[1..n]` to `[0..1]`.
}
\details{
[qtile.integer64()] is the inverse function of 'prank.integer64' and
  projects `[0..1]` to `[min..max]`.

Function `prank.integer64` is based on [rank.integer64()].
}
\examples{
x <- as.integer64(sample(c(rep(NA, 9), 1:9), 32, TRUE))
prank(x)

x <- x[!is.na(x)]
stopifnot(identical(x,  unname(qtile(x, probs=prank(x)))))
}
\seealso{
[rank.integer64()] for simple ranks and [qtile()] for the
  inverse function quantiles.
}
\keyword{univar}
