% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integer64.R
\name{seq.integer64}
\alias{seq.integer64}
\title{integer64: Sequence Generation}
\arguments{
\item{from}{integer64 scalar (in order to dispatch the integer64 method of [seq()]}

\item{to}{scalar}

\item{by}{scalar}

\item{length.out}{scalar}

\item{along.with}{scalar}

\item{...}{ignored}
}
\value{
an integer64 vector with the generated sequence
}
\description{
Generating sequence of integer64 values
}
\details{
`seq.integer64` does coerce its arguments 'from', 'to' and 'by' to `integer64`.
  If not provided, the argument 'by' is automatically determined as `+1` or `-1`,
  but the size of 'by' is not calculated as in [seq()] (because this might result
  in a non-integer value).
}
\note{
In base R [`:`] currently is not generic and does not dispatch, see section
  "Limitations inherited from Base R" in [integer64()]
}
\examples{
  # colon not activated: as.integer64(1):12
  seq(as.integer64(1), 12, 2)
  seq(as.integer64(1), by=2, length.out=6)
}
\seealso{
[c.integer64()] [rep.integer64()]
  [as.data.frame.integer64()] [integer64()]
}
\keyword{classes}
\keyword{manip}
